demo/edge-services/edge-service/requirements.txt
flask==3.0.3
httpx==0.27.0
Pillow==10.3.0
depthai
